# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/teams": {
      "post": {
        "operationId": "AdminTeamsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamWriteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TeamIdResponseDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Takim olustur",
        "tags": [
          "admin-teams"
        ]
      }
    },
    "/api/admin/teams/{id}": {
      "put": {
        "operationId": "AdminTeamsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamWriteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Team not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Takim guncelle",
        "tags": [
          "admin-teams"
        ]
      },
      "delete": {
        "operationId": "AdminTeamsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Team not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Takim sil",
        "tags": [
          "admin-teams"
        ]
      }
    },
...
    "/api/profile/photo": {
      "post": {
        "operationId": "ProfilePhotoController_upload",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProfilePhotoUrlDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Profil fotoğrafı yükle",
        "tags": [
          "profile"
        ]
      },
      "put": {
        "operationId": "ProfilePhotoController_replace",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProfilePhotoUrlDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Profil fotoğrafı değiştir",
        "tags": [
          "profile"
        ]
      },
      "get": {
        "operationId": "ProfilePhotoController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProfilePhotoUrlDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Profil fotoğrafını getir",
        "tags": [
          "profile"
        ]
      },
      "delete": {
        "operationId": "ProfilePhotoController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Profil fotoğrafını sil",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/photo/from-url": {
      "post": {
        "operationId": "ProfilePhotoController_fromUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageUrlDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProfilePhotoUrlDto"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "URL’den profil fotoğrafı",
        "tags": [
          "profile"
        ]
      }
    },
...
  },
  "info": {
    "title": "TransferPulse API",
    "description": "TransferPulse backend (.NET → NestJS migration)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      },
...
      "ImageUrlDto": {
        "type": "object",
        "properties": {
          "imageUrl": {
            "type": "string",
            "description": "İndirilecek görsel URL"
          }
        },
        "required": [
          "imageUrl"
        ]
      },
...
      "TeamIdResponseDto": {
        "type": "object",
        "properties": {
          "teamId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "required": [
          "teamId"
        ]
      },
      "TeamWriteDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50
          },
          "logo": {
            "type": "string"
          },
          "leagueId": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "leagueId"
        ]
      },
...
      "ProfilePhotoUrlDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://cdn.example.com/photos/abc.jpg"
          }
        },
        "required": [
          "url"
        ]
      }
    }
  }
}